requirements.txt
flask_sqlalchemy
sqlalchemy
pydantic_sqlalchemy
flask